+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - [INTERNAL] fix npm security alerts
-- Update textile-js to , to get HTML comments line-numbering, and a fix for `<LI>`
+- Update textile-js to GehDoc/textile-js@72428b65a3e14fb2ebb370d845786790aeb3e7f8, to get HTML comments line-numbering, and a fix for `<LI>`
 
 ### Changed
 - [INTERNAL] Include shared.webpack config to diffs.