+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [DOC] Add SECURITY.md
 
+### Fixed
+- [INTERNAL] npm audit fix
+
 
 ## [1.2.0] - 20200323
 ### Changed