+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [UNRELEASED]
+## [1.2.1] - 20200329
 ### Added
 - [DOC] Add SECURITY.md
 - [PREVIEW] Add paragraph folding for multi-line `<LI>` and HTML code blocks (PRE, DIV, BLOCKQUOTE)